Added a ref directive to the input field.
A ref directive was added to the input field to allow it to be referenced
from outside the component, for example to set input focus. If a ref
directive is added to the v-autocomplete component you can reference the
input field inside the component like this:

    this.$refs['autocomplete'].$refs['input'].focus()

where 'autocomplete' is the component ref and 'input' is the input field
ref. A name of 'input' was used since it's obvious what it refers to and
there are not other refs in the component thta it might clash with.